{{cookiecutter.github_repository_name}}/requirements.txt
dj-database-url==0.5.0
Django==1.11.7
gunicorn==19.8.1
psycopg2==2.7.5
python-decouple==3.1
redis==2.10.6
urlwait==0.4

{%- if cookiecutter.use_celery == "y" %}
# Celery requirements
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
kombu==4.2.1
pytz==2018.5
vine==1.1.4
{%- endif %}

{%- if cookiecutter.use_sentry == "y" %}
# Using raven for sentry
raven==6.7.0
{%- endif %}

{%- if cookiecutter.use_rest_framework == "y" %}
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11

{%- if "{{ cookiecutter.email_user }}".lower() == "n" %}
factory-boy==2.10.0

pytest-django==3.1.2
pytest-cov==2.5.1
{%- endif %}

{%- endif %}

{%- if cookiecutter.use_rest_framework == "y" and cookiecutter.use_rest_auth == "y" %}
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
django-allauth==0.35.0
django-rest-auth==0.9.3
djangorestframework-jwt==1.11.0
idna==2.7
oauthlib==2.0.7
PyJWT==1.6.1
python3-openid==3.1.0
{%- if cookiecutter.use_celery == "n" %}
pytz==2018.5
{%- endif %}
requests==2.19.1
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.23
{%- endif %}

{%- if cookiecutter.use_rest_framework == "y" and cookiecutter.use_restframework_documentation == "y" %}
{%- if cookiecutter.use_rest_framework == "y" and cookiecutter.use_rest_auth == "n" %}
certifi==2018.4.16
chardet==3.0.4
idna==2.7
requests==2.19.1
urllib3==1.23
{%- endif %}
coreapi==2.3.3
coreschema==0.0.4
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
uritemplate==3.0.0
{%- endif %}
